% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_final_weights.R
\name{calculate_final_weights}
\alias{calculate_final_weights}
\title{Title}
\usage{
calculate_final_weights(
  gene_list,
  protein_file,
  mirna_file,
  use_full_data = TRUE
)
}
\arguments{
\item{gene_list}{The gene list for analysis.}

\item{protein_file}{A total file of protein relationship.}

\item{mirna_file}{The miRNA-mRNA relationship file to obtain vital miRNA.}

\item{use_full_data}{The selection to determine whether to use all proteins.}
}
\value{
Final_weights
}
\description{
To calculate genes weights based on protein and miRNA relation.
}
\examples{
library(data.table)
library(igraph)
library(biomaRt)
library(ggraph)
library(R.utils)# #
protein_file <- system.file("extdata", "9606.protein.links.v12.0.txt.gz", package = "XYlabmiR")
mirna_file <- system.file("extdata", "miRTarget.zip", package = "XYlabmiR")#
#
gene_list <- c("EGFR", "EPCAM", "VIM")#
calculate_final_weights(gene_list, protein_file, mirna_file, use_full_data = TRUE)
}
