% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_node_importance.R
\name{calculate_node_importance}
\alias{calculate_node_importance}
\title{Title}
\usage{
calculate_node_importance(gene_list, protein_file, use_full_data = TRUE)
}
\arguments{
\item{gene_list}{The gene list for analysis.}

\item{protein_file}{A total file of protein relationship.}

\item{use_full_data}{The miRNA-mRNA relationship file to obtain vital miRNA.}
}
\value{
A result containing genes importance.
}
\description{
To obtain aim genes.
}
\examples{
library(data.table)
library(igraph)
library(biomaRt)
library(ggraph)
library(R.utils)
#
protein_file <- system.file("extdata", "9606.protein.links.v12.0.txt.gz", package = "XYlabmiR")

#
gene_list <- c("EGFR", "EPCAM", "VIM")

calculate_node_importance(gene_list, protein_file, use_full_data = TRUE)
}
