% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mirna_gene_relationship.R
\name{get_mirna_gene_relationship}
\alias{get_mirna_gene_relationship}
\title{Title}
\usage{
get_mirna_gene_relationship(mirna_file)
}
\arguments{
\item{mirna_file}{The miRNA-mRNA relationship file to obtain vital miRNAs.}
}
\value{
The result of important miRNAs counts.
}
\description{
Title
}
\examples{
library(data.table)
library(igraph)
library(biomaRt)
library(ggraph)
library(R.utils)
#
mirna_file <- system.file("extdata", "miRTarget.zip", package = "XYlabmiR")

#
gene_list <- c("EGFR", "EPCAM", "VIM")
get_mirna_gene_relationship(mirna_file)
}
